fix(log_view): move MedView.tsx to page.tsx
Turns out, MedView.tsx is a newer version of page.tsx. page.tsx
was deleted before, but this reverts it to prevent a 404 after
opening the Medication Log view.
